requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.7
colorama==0.4.6
Django==4.2.2
django-cors-headers==4.1.0
django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.6
gunicorn==20.1.0
inflection==0.5.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==9.5.0
platformdirs==3.10.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.5.0
